random grid position
Added code to start the town at a random position on the underlying grid. This stops us form getting the same recognisable features in the centre of every town generated.

Instead of setting up min/max x and y coordinates in the .ini file, we simplay specify the size of the town and this is added to a random integer in the main generator script.